web_ui/requirements.txt
Django==1.10
google-api-python-client==1.5.3
httplib2==0.9.2
msgpack-numpy==0.3.7
msgpack-python==0.4.8
numpy==1.11.1
oauth2client==3.0.0
pytz==2016.6.1
ujson==1.35
